1.3.6.5 version. + fix for wool placing near points bug.

+++ src/me/dalton/capturethepoints/Items.java
@@ -12,5 +12,5 @@ public class Items {
     public int amount;
     
     /** The items data type */
-    public int type = -1;
+    public short type = -1;
 }
